Added GetLanguageCount endpoint. Renamed GetLanguageCount endpoint path to better match others
