+++ Logic/Endpoints/Languages/GetLanguageCount/GetLanguageCountResponse.cs
@@ -1,0 +1,7 @@
+namespace CourseManagement.Logic.Endpoints.Languages.GetLanguageCount
+{
+    public class GetLanguageCountResponse
+    {
+        public int Count { get; set; }
+    }
+}
