+++ Logic/Endpoints/Languages/GetLanguageList/GetLanguageListEndpoint.cs
@@ -3,7 +3,7 @@ using FastEndpoints;
 
 namespace CourseManagement.Logic.Endpoints.Languages.GetLanguageList
 {
-    public class GetLanguageListEndpoint : Endpoint<EmptyRequest, List<GetLanguageListResponse>, GetLanguageListMapper>
+    public class GetLanguageListEndpoint : Endpoint<GetLanguageListRequest, List<GetLanguageListResponse>, GetLanguageListMapper>
     {
         public override void Configure()
         {
@@ -18,9 +18,9 @@ namespace CourseManagement.Logic.Endpoints.Languages.GetLanguageList
             this.courseDbContext = courseDbContext;
         }
 
-        public override async Task HandleAsync(EmptyRequest req, CancellationToken ct)
+        public override async Task HandleAsync(GetLanguageListRequest req, CancellationToken ct)
         {
-            List<Data.Models.Language> languages = courseDbContext.Languages.ToList();
+            List<Data.Models.Language> languages = courseDbContext.Languages.Skip(req.Skip).Take(req.Take).ToList();
             Response = Map.FromEntity(languages);
             await SendOkAsync(Response, ct);
         }
