+++ Logic/Endpoints/Skills/GetSkillList/GetSkillListRequest.cs
@@ -2,7 +2,7 @@ namespace CourseManagement.Logic.Endpoints.Skills.GetSkillList
 {
     public class GetSkillListRequest
     {
-        public required int Skip { get; set; }
-        public required int Take { get; set; }
+        public required int Skip { get; set; } = 0;
+        public required int Take { get; set; } = 10;
     }
 }
